requirements.txt
Werkzeug
numpy
flask
torch
torchvision
Pillow
